Fix &va_list on x86-64, quiet warning
UltraBlame original commit: 6013a86204c2f5bd64b74620f694fc3636cb5498

+++ js/src/jshashtable.h
@@ -242,7 +242,7 @@ class HashTable : AllocPolicy
 #endif
 
     static const unsigned sMinSize      = 16;
-    static const unsigned sSizeLimit    = tl::NBitMask<24>::result;
+    static const size_t   sSizeLimit    = tl::NBitMask<24>::result;
     static const unsigned sHashBits     = tl::BitSize<HashNumber>::result;
     static const unsigned sGoldenRatio  = 0x9E3779B9U;       
     static const uint8    sMinAlphaFrac = 64;
